+++ ios/THEOplayerRCTMainEventHandler.swift
@@ -29,7 +29,8 @@ public class THEOplayerRCTMainEventHandler {
     var onNativeRateChange: RCTDirectEventBlock?
     var onNativeWaiting: RCTDirectEventBlock?
     var onNativeCanPlay: RCTDirectEventBlock?
-    var onNativeResize: RCTDirectEventBlock?
+    var onNativeDimensionChange: RCTDirectEventBlock?
+    var onNativeVideoResize: RCTDirectEventBlock?
     
     // MARK: player Listeners
     private var playListener: EventListener?
@@ -51,11 +52,10 @@ public class THEOplayerRCTMainEventHandler {
     private var rateChangeListener: EventListener?
     private var waitingListener: EventListener?
     private var canPlayListener: EventListener?
-    private var resizeListener: EventListener?
-    private var presentationModeChangeListener: EventListener?
+    private var videoResizeListener: EventListener?
     
     // MARK: player observer
-    private var videoRectObserver: NSKeyValueObservation?
+    private var dimensionChangeObserver: NSKeyValueObservation?
     
     // MARK: - destruction
     func destroy() {
@@ -310,27 +310,27 @@ public class THEOplayerRCTMainEventHandler {
         }
         if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] CanPlay listener attached to THEOplayer") }
         
-        // RESIZE: TODO on iOS SDK
-        /*self.resizeListener = player.addEventListener(type: PlayerEventTypes.RESIZE) { [weak self, weak player] event in
+        // RESIZE
+        self.videoResizeListener = player.addEventListener(type: PlayerEventTypes.RESIZE) { [weak self, weak player] event in
             if DEBUG_THEOPLAYER_EVENTS { PrintUtils.printLog(logText: "[NATIVE] Received RESIZE event from THEOplayer") }
             if let wplayer = player,
-               let forwardedResizeEvent = self?.onNativeResize {
-                forwardedResizeEvent(
+               let forwardedVideoResizeEvent = self?.onNativeVideoResize {
+                forwardedVideoResizeEvent(
                     [
-                        "width": wplayer.frame.width,
-                        "height": wplayer.frame.height,
+                        "videoWidth": wplayer.videoWidth,
+                        "videoHeight": wplayer.videoHeight,
                     ]
                 )
             }
         }
-        if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] Resize listener attached to THEOplayer") }*/
+        if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] Resize listener attached to THEOplayer") }
         
-        // RESIZE: implemented using videoRect Observation
-        self.videoRectObserver = player.observe(\.videoRect, options: [.new]) { [weak self, weak player] view, change in
+        // DIMENSION CHANGE: implemented using videoRect Observation
+        self.dimensionChangeObserver = player.observe(\.videoRect, options: [.new]) { [weak self, weak player] view, change in
             if DEBUG_THEOPLAYER_EVENTS { PrintUtils.printLog(logText: "[NATIVE] Observed videoRect change on THEOplayer") }
             if let wplayer = player,
-               let forwardedResizeEvent = self?.onNativeResize {
-                forwardedResizeEvent(
+               let forwardedDimensionChangeEvent = self?.onNativeDimensionChange {
+                forwardedDimensionChangeEvent(
                     [
                         "width": wplayer.frame.width,
                         "height": wplayer.frame.height,
@@ -453,12 +453,15 @@ public class THEOplayerRCTMainEventHandler {
             if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] CanPlay listener dettached from THEOplayer") }
         }
         
+        // DIMENSION CHANGE
+        if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] Invalidate dimensionChangeObserver from THEOplayer") }
+        self.dimensionChangeObserver?.invalidate()
+        self.dimensionChangeObserver = nil
+      
         // RESIZE
-        /*if let resizeListener = self.resizeListener {
-            player.removeEventListener(type: PlayerEventTypes.RESIZE, listener: resizeListener)
-            if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] Resize listener dettached from THEOplayer") }
-        }*/
-        self.videoRectObserver?.invalidate()
-        self.videoRectObserver = nil
+        if let videoResizeListener = self.videoResizeListener {
+          player.removeEventListener(type: PlayerEventTypes.RESIZE, listener: videoResizeListener)
+          if DEBUG_EVENTHANDLER { PrintUtils.printLog(logText: "[NATIVE] Resize listener dettached from THEOplayer") }
+        }
     }
 }