+++ src/internal/adapter/event/PlayerEvents.ts
@@ -50,6 +50,8 @@ import {
   TheoLiveEndpoint,
   TheoLiveEndpointLoadedEvent,
   TheoLiveIntentToFallbackEvent,
+  DimensionChangeEvent,
+  VideoResizeEvent,
 } from 'react-native-theoplayer';
 
 export class DefaultLoadedMetadataEvent extends BaseEvent<PlayerEventType.LOADED_METADATA> implements LoadedMetadataEvent {
@@ -91,12 +93,33 @@ export class DefaultVolumeChangeEvent extends BaseEvent<PlayerEventType.VOLUME_C
   }
 }
 
+/**
+ * @deprecated Use {@link DefaultDimensionChangeEvent} instead. This event is set for removal in version 11.
+ */
 export class DefaultResizeEvent extends BaseEvent<PlayerEventType.RESIZE> implements ResizeEvent {
   constructor(
     public width: number,
     public height: number,
   ) {
     super(PlayerEventType.RESIZE);
+  }
+}
+
+export class DefaultDimensionChangeEvent extends BaseEvent<PlayerEventType.DIMENSION_CHANGE> implements DimensionChangeEvent {
+  constructor(
+    public width: number,
+    public height: number,
+  ) {
+    super(PlayerEventType.DIMENSION_CHANGE);
+  }
+}
+
+export class DefaultVideoResizeEvent extends BaseEvent<PlayerEventType.VIDEO_RESIZE> implements VideoResizeEvent {
+  constructor(
+    public videoWidth: number,
+    public videoHeight: number,
+  ) {
+    super(PlayerEventType.VIDEO_RESIZE);
   }
 }