+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [10.1.0] - 25-10-06
+
+### Added
+
+- Added an explicit clearing of the NowPlayingInfo on iOS, when the app receives a `willTerminateNotification`, to make sure all NowPlayingInfo is removed from the lock screen when an app is closed.
+- Added `videoWidth` and `videoHeight` properties to `THEOplayer` containing the active video's resolution in pixels.
+- Added `videoresize` player event enabling monitoring of video rendition resizes.
+- Added `dimensionchange` player event enabling monitoring of player resizes. This event is set to replace the now deprecated `resize` player event.
+- Added `fontPath` to `TextTrackStyle` to allow loading custom fonts from the Android assets folder.
+
+### Fixed
+
+- Fixed an issue on Android where play-out of Millicast sources in the example app would fail. The Android NDK version needs to be at least v28.
+- Fixed an issue on Web where on older devices the `display: contents` style property is sometimes not supported.
+- Fixed an issue on Android where the `MediaPlaybackService` could be restarted after the app was closed, in some cases causing a `ForegroundServiceDidNotStartInTimeException`. A conditional receiver makes sure the service is never restarted.
+- Fixed an issue on iOS where a setup with multiple players could cause a view index NSInternalInconsistencyException after moving players in and out of fullscreen.
+- Fixed an issue on Android where some event listeners were not properly disposed of after destroying the player.
+- Fixed an issue on Web where the `width` and `height` properties on `THEOplayer` would show the active video rendition's resolution width and height instead of the width and height of the video element.
+
 ## [10.0.0] - 25-09-15
 
 ### Added